% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_disease_names.R
\name{check_disease_names}
\alias{check_disease_names}
\title{Check if the \code{disease_names} is valid.}
\usage{
check_disease_names(disease_names)
}
\arguments{
\item{disease_names}{disease names.
Possible value is "healthy"}
}
\value{
nothing
}
\description{
Check if the \code{disease_names} is valid.
Will \link{stop} if not
}
\examples{
check_disease_names("cs.{healthy}")
}
\author{
Richèl J.C. Bilderbeek
}
