% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_error_query_results.R
\name{check_error_query_results}
\alias{check_error_query_results}
\title{Check that the \code{query_results} is a success result}
\usage{
check_error_query_results(query_results)
}
\arguments{
\item{query_results}{the results of a query
(e.g. a call to \link{query_iedb}) as a \link[tibble]{tibble}}
}
\value{
Nothing. Will stop if \code{query_results} is an error result.
}
\description{
Check that the \code{query_results} is a success result,
i.e. that the query did result in a useful lookup from the database
}
\author{
Richèl J.C. Bilderbeek
}
