% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_host_organism_iris.R
\name{check_host_organism_iris}
\alias{check_host_organism_iris}
\title{Check if the \code{host_organism_iris} is valid.}
\usage{
check_host_organism_iris(host_organism_iris)
}
\arguments{
\item{host_organism_iris}{the host.
Possible value is "NCBITaxon:9606", which denotes Homo sapiens}
}
\value{
nothing
}
\description{
Check if the \code{host_organism_iris} is valid.
Will \link{stop} if not
}
\examples{
check_host_organism_iris("cs.{NCBITaxon:9606}")
}
\author{
Richèl J.C. Bilderbeek
}
