% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_max_n_queries.R
\name{check_max_n_queries}
\alias{check_max_n_queries}
\title{Check if the \code{max_n_queries} is valid.}
\usage{
check_max_n_queries(max_n_queries)
}
\arguments{
\item{max_n_queries}{maximum number of queries.
Use \link{Inf} to have no upper limit on this.
Lower values are useful for debugging.}
}
\value{
nothing
}
\description{
Check if the \code{max_n_queries} is valid.
Will \link{stop} if not
}
\examples{
check_max_n_queries(max_n_queries = Inf)
check_max_n_queries(max_n_queries = 3)
}
\author{
Richèl J.C. Bilderbeek
}
