% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhc_allele_name.R
\name{check_mhc_allele_name}
\alias{check_mhc_allele_name}
\title{Check if the \code{mhc_allele_name} is valid.}
\usage{
check_mhc_allele_name(mhc_allele_name, mhc_allele_name_regex = "HLA-.*")
}
\arguments{
\item{mhc_allele_name}{name of an MHC allele}

\item{mhc_allele_name_regex}{a regular expression for
an MHC allele names}
}
\value{
nothing
}
\description{
Check if the \code{mhc_allele_name} is valid.
Will \link{stop} if not
}
\examples{
check_mhc_allele_name("HLA-A*01:01")
check_mhc_allele_name("all")
}
\seealso{
use \link{check_mhc_allele_names} to check a
selection of one or more alleles, as based on their names
}
\author{
Richèl J.C. Bilderbeek
}
