% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_order.R
\name{check_order}
\alias{check_order}
\title{Check if the \code{order} is valid.}
\usage{
check_order(order)
}
\arguments{
\item{order}{the item by which the query is ordered,
as checked by \link{check_order}.
For bigger queries (i.e. with more than 10k results),
\code{order} is important to obtain each 10k set of results in
the same manner.}
}
\value{
nothing
}
\description{
Check if the \code{order} is valid.
Will \link{stop} if not
}
\examples{
check_order("structure_iri")
}
\author{
Richèl J.C. Bilderbeek
}
