% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_query.R
\name{check_query}
\alias{check_query}
\title{Check if the \code{query} is valid.}
\usage{
check_query(query)
}
\arguments{
\item{query}{an IEDB query,
as created by, for example, \link{create_healthy_human_query}.
Use \link{check_query} to check if the \code{query} is valid}
}
\value{
nothing
}
\description{
Check if the \code{query} is valid.
Will \link{stop} if not
}
\examples{
check_query("structure_iri")
}
\author{
Richèl J.C. Bilderbeek
}
