% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_select.R
\name{check_select}
\alias{check_select}
\title{Check if the \code{select} is valid.}
\usage{
check_select(select)
}
\arguments{
\item{select}{the fields that the query returns,
as checked by \link{check_select}}
}
\value{
nothing
}
\description{
Check if the \code{select} is valid.
Will \link{stop} if not
}
\examples{
check_select("linear_sequence")
}
\author{
Richèl J.C. Bilderbeek
}
