% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_source_organism_iris.R
\name{check_source_organism_iris}
\alias{check_source_organism_iris}
\title{Check if the \code{source_organism_iris} is valid.}
\usage{
check_source_organism_iris(source_organism_iris)
}
\arguments{
\item{source_organism_iris}{the epitope source.
Possible value is "NCBITaxon:9606", which denotes Homo sapiens}
}
\value{
nothing
}
\description{
Check if the \code{source_organism_iris} is valid.
Will \link{stop} if not
}
\examples{
check_source_organism_iris("NCBITaxon:9606")
}
\author{
Richèl J.C. Bilderbeek
}
