% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure_type.R
\name{check_structure_type}
\alias{check_structure_type}
\title{Check if the \code{structure_type} is valid.}
\usage{
check_structure_type(structure_type)
}
\arguments{
\item{structure_type}{structure of the epitope,
as checked by \link{check_structure_type}.
Possible values are "Linear peptide",
as these are the names used at \url{http://www.iedb.org/}}
}
\value{
nothing
}
\description{
Check if the \code{structure_type} is valid.
Will \link{stop} if not
}
\examples{
check_structure_type("eq.Linear peptide")
}
\author{
Richèl J.C. Bilderbeek
}
