% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table.R
\name{check_table}
\alias{check_table}
\title{Check if the \code{table} is valid.}
\usage{
check_table(table)
}
\arguments{
\item{table}{a IEDB database table,
as checked by \link{check_table}}
}
\value{
nothing
}
\description{
Check if the \code{table} is valid.
Will \link{stop} if not
}
\examples{
check_table("mhc_search")
check_table("epitope_search")
}
\author{
Richèl J.C. Bilderbeek
}
