% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_verbose.R
\name{check_verbose}
\alias{check_verbose}
\title{Check if the \code{verbose} is valid.}
\usage{
check_verbose(verbose)
}
\arguments{
\item{verbose}{set to \link{TRUE} for more output}
}
\value{
nothing
}
\description{
Check if the \code{verbose} is valid.
Will \link{stop} if not
}
\examples{
check_verbose(TRUE)
check_verbose(FALSE)
}
\author{
Richèl J.C. Bilderbeek
}
