% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_healthy_human_query.R
\name{create_healthy_human_query}
\alias{create_healthy_human_query}
\title{Create a query for the IEDB for a linear peptide in a healthy human.}
\usage{
create_healthy_human_query(
  structure_type = "eq.Linear peptide",
  source_organism_iris = "cs.{NCBITaxon:9606}",
  host_organism_iris = "cs.{NCBITaxon:9606}",
  disease_names = "cs.{healthy}",
  order = "structure_iri"
)
}
\arguments{
\item{structure_type}{structure of the epitope,
as checked by \link{check_structure_type}.
Possible values are "Linear peptide",
as these are the names used at \url{http://www.iedb.org/}}

\item{source_organism_iris}{the epitope source.
Possible value is "NCBITaxon:9606", which denotes Homo sapiens}

\item{host_organism_iris}{the host.
Possible value is "NCBITaxon:9606", which denotes Homo sapiens}

\item{disease_names}{disease names.
Possible value is "healthy"}

\item{order}{the item by which the query is ordered,
as checked by \link{check_order}.
For bigger queries (i.e. with more than 10k results),
\code{order} is important to obtain each 10k set of results in
the same manner.}
}
\description{
Create a query for the IEDB for a linear peptide in a healthy human.
The parameters are in the same order as \url{http://www.iedb.org/}
}
\author{
Richèl J.C. Bilderbeek
}
