% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  assays,
  disease_names,
  mhc_restriction,
  source_organism_iris,
  structure_type,
  verbose
)
}
\arguments{
\item{assays}{one or more assays.
Possible values are "B Cell", "MHC Ligand" and "T Cell",
as these are the names used at \url{http://www.iedb.org/}}

\item{disease_names}{disease names.
Possible value is "healthy"}

\item{mhc_restriction}{MHC restriction
Possible values are "Any", "Class I",
"Class II", "Non-classical" and a haplotype name}

\item{source_organism_iris}{the epitope source.
Possible value is "NCBITaxon:9606", which denotes Homo sapiens}

\item{structure_type}{structure of the epitope,
as checked by \link{check_structure_type}.
Possible values are "Linear peptide",
as these are the names used at \url{http://www.iedb.org/}}

\item{verbose}{set to TRUE for more output}

\item{host_organism_iris}{the host.
Possible value is "NCBITaxon:9606", which denotes Homo sapiens}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
