% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_b_or_t_cell_epitopes_and_mhc_alleles.R
\name{get_all_b_or_t_cell_epitopes_and_mhc_alleles}
\alias{get_all_b_or_t_cell_epitopes_and_mhc_alleles}
\title{Internal function}
\usage{
get_all_b_or_t_cell_epitopes_and_mhc_alleles(
  b_cell_or_t_cell,
  max_n_queries = Inf,
  verbose = FALSE
)
}
\arguments{
\item{b_cell_or_t_cell}{use "b_cell" for B cells
and "t_cell" for T cells}

\item{max_n_queries}{maximum number of queries.
Use \link{Inf} to have no upper limit on this.
Lower values are useful for debugging.}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
a \link[tibble]{tibble} tibble with two columns:
\itemize{
\item \code{linear_sequence}: the linear sequence of the epitope
\item \code{mhc_allele_name}: the name of the MHC allele
}
}
\description{
Get all epitopes from a B or T cell assay
}
\examples{
if (pureseqtmr::is_on_ci()) {
  get_all_b_or_t_cell_epitopes_and_mhc_alleles(b_cell_or_t_cell = "b_cell")
  get_all_b_or_t_cell_epitopes_and_mhc_alleles(b_cell_or_t_cell = "t_cell")
}
}
\author{
Richèl J.C. Bilderbeek
}
