% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_mhc_ligand_epitopes.R
\name{get_all_mhc_ligand_epitopes}
\alias{get_all_mhc_ligand_epitopes}
\title{Get all epitopes from an MHC ligand assay}
\usage{
get_all_mhc_ligand_epitopes(verbose = FALSE)
}
\arguments{
\item{verbose}{set to \link{TRUE} for more output}
}
\value{
all epitopes
}
\description{
Get all epitopes from an MHC ligand assay
}
\examples{
if (beastier::is_on_ci()) {
  get_all_b_cell_epitopes()
}
}
\author{
Richèl J.C. Bilderbeek
}
