% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_mhc_ligand_epitopes_and_mhc_alleles.R
\name{get_all_mhc_ligand_epitopes_and_mhc_alleles}
\alias{get_all_mhc_ligand_epitopes_and_mhc_alleles}
\title{Get all epitopes with MHC alleles from an MHC ligand assay}
\usage{
get_all_mhc_ligand_epitopes_and_mhc_alleles(
  max_n_queries = Inf,
  verbose = FALSE
)
}
\arguments{
\item{max_n_queries}{maximum number of queries.
Use \link{Inf} to have no upper limit on this.
Lower values are useful for debugging.}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
a \link[tibble]{tibble} tibble with two columns:
\itemize{
\item \code{linear_sequence}: the linear sequence of the epitope
\item \code{mhc_allele_name}: the name of the MHC allele
}
}
\description{
Get all epitopes with MHC alleles from an MHC ligand assay
}
\examples{
if (pureseqtmr::is_on_ci()) {
  get_all_mhc_ligand_epitopes_and_mhc_alleles(
    max_n_queries = 3
  )
}
}
\seealso{
\itemize{
\item use \link{get_all_b_cell_epitopes_and_mhc_alleles}
to get all epitopes from a B cell assay
\item use \link{get_all_mhc_ligand_epitopes_and_mhc_alleles}
to get all epitopes from an MHC ligand assay
\item use \link{get_all_t_cell_epitopes_and_mhc_alleles}
to get all epitopes from a T cell assay
}
}
\author{
Richèl J.C. Bilderbeek
}
