% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_t_cell_epitopes.R
\name{get_all_t_cell_epitopes}
\alias{get_all_t_cell_epitopes}
\title{Get all epitopes from a B cell assay}
\usage{
get_all_t_cell_epitopes(
  mhc_allele_names = "all",
  max_n_queries = Inf,
  verbose = FALSE
)
}
\arguments{
\item{mhc_allele_names}{names of one or more MHC allele names}

\item{max_n_queries}{maximum number of queries.
Use \link{Inf} to have no upper limit on this.
Lower values are useful for debugging.}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
all epitopes
}
\description{
Get all epitopes from a B cell assay
}
\examples{
if (pureseqtmr::is_on_ci()) {
  get_all_t_cell_epitopes()
}
}
\seealso{
\itemize{
\item use \link{get_all_b_cell_epitopes}
to get all epitopes from a B cell assay
\item use \link{get_all_mhc_ligand_epitopes}
to get all epitopes from an MHC ligand assay
\item use \link{get_all_t_cell_epitopes}
to get all epitopes from a T cell assay
}
}
\author{
Richèl J.C. Bilderbeek
}
