% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selectors.R
\name{get_selectors}
\alias{get_selectors}
\title{Get all valid values for \code{select}}
\usage{
get_selectors()
}
\value{
all valid values for \code{select}
}
\description{
Get all valid values for \code{select},
as checked by \link{check_select}
}
\examples{
get_selectors()
}
\author{
Richèl J.C. Bilderbeek
}
