% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_query_results_error.R
\name{get_test_query_results_error}
\alias{get_test_query_results_error}
\title{Get a \code{query_results} that is an error result, to be used in testing}
\usage{
get_test_query_results_error()
}
\value{
a \code{query_results} that is an error result
}
\description{
Get a \code{query_results} that is an error result, to be used in testing
}
\author{
Richèl J.C. Bilderbeek
}
