% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_error_query_results.R
\name{is_error_query_results}
\alias{is_error_query_results}
\title{Determine if the \code{query_results} is an error result}
\usage{
is_error_query_results(query_results)
}
\arguments{
\item{query_results}{the results of a query
(e.g. a call to \link{query_iedb}) as a \link[tibble]{tibble}}
}
\value{
\itemize{
\item \link{TRUE} if the \code{query_results} is an error result.
\item \link{FALSE} if the \code{query_results} is not error result.
}

Will \link{stop} if \code{query_results} is invalid
}
\description{
Determine if the \code{query_results} is an error result,
i.e. that the query did not result in a lookup from the datebase
and contains an error message (and hint) instead.
}
\examples{
is_error_query_results(get_test_query_results_error())
}
\author{
Richèl J.C. Bilderbeek
}
