% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhc_allele_name_to_regex.R
\name{mhc_allele_name_to_regex}
\alias{mhc_allele_name_to_regex}
\title{Convert an \code{mhc_allele_name} to a regular expression to detect it}
\usage{
mhc_allele_name_to_regex(mhc_allele_name)
}
\arguments{
\item{mhc_allele_name}{name of an MHC allele}
}
\value{
the regular expression to detect that \code{mhc_allele_name}
}
\description{
Convert an \code{mhc_allele_name} to a regular expression to detect it
}
\examples{
mhc_allele_name <- "HLA-DRB*01:01"
mhc_allele_name_regex <- mhc_allele_name_to_regex(mhc_allele_name)

# Can detect the regex
stringr::str_detect(
  string = "HLA-DRB*01:01",
  pattern = mhc_allele_name_regex
)

# Will not detect a match to the regex
stringr::str_detect(
  string = "HLA-DRB.01:01", # replaced asterisk by dot
  pattern = mhc_allele_name_regex
)
}
\author{
Richèl J.C. Bilderbeek
}
