% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_iedb.R
\name{query_iedb}
\alias{query_iedb}
\title{Query IEDB}
\usage{
query_iedb(query, verbose = FALSE)
}
\arguments{
\item{verbose}{set to \link{TRUE} for more output}
}
\description{
Query IEDB
}
\author{
Richèl J.C. Bilderbeek
}
