% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_iedb_with_offset.R
\name{query_iedb_with_offset}
\alias{query_iedb_with_offset}
\title{Query the IEDB database from an offset}
\usage{
query_iedb_with_offset(query, table, offset = 0, verbose = FALSE)
}
\arguments{
\item{query}{an IEDB query,
as created by, for example, \link{create_healthy_human_query}.
Use \link{check_query} to check if the \code{query} is valid}

\item{offset}{the offset of the results from a query,
as checked by \link{check_offset}.
A standard query uses an offset of zero
and returns maximally \code{10000} results.
To obtain the next results, an \code{offset} of \code{10000}
is used.}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
a \link[tibble]{tibble} with as much rows as hits.
}
\description{
Query the IEDB database from an offset.
}
\note{
The IEDB returns 10k hits per query.
This means, to get the results of a query of more than 10k hits,
the IEDB needs to be queried multiple times.
\link{query_iedb} combines multiple calls to the IEDB,
using \link{query_iedb_with_offset}.
}
\author{
Richèl J.C. Bilderbeek
}
