% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mhc_allele_names.R
\name{check_mhc_allele_names}
\alias{check_mhc_allele_names}
\title{Check if the \code{mhc_allele_names} is valid.}
\usage{
check_mhc_allele_names(
  mhc_allele_names,
  mhc_allele_names_regex = "(cs)\\\\.\\\\{(HLA-.*)\\\\}"
)
}
\arguments{
\item{mhc_allele_names}{names of one or more MHC allele names}

\item{mhc_allele_names_regex}{a regular expression for
a 'PostgREST' expression to select zero, one or more MHC allele names}
}
\value{
nothing
}
\description{
Check if the \code{mhc_allele_names} is valid.
Will \link{stop} if not
}
\examples{
check_mhc_allele_names("cs.{HLA-A*01:01}")
}
\seealso{
use \link{check_mhc_allele_name} to check an allele name
}
\author{
Richèl J.C. Bilderbeek
}
