% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_offset.R
\name{check_offset}
\alias{check_offset}
\title{Check if the \code{offset} is valid.}
\usage{
check_offset(offset)
}
\arguments{
\item{offset}{the offset of the results from a query,
as checked by \link{check_offset}.
A standard query uses an offset of zero
and returns maximally \code{10000} results.
To obtain the next results, an \code{offset} of \code{10000}
is used.}
}
\value{
nothing
}
\description{
Check if the \code{offset} is valid.
Will \link{stop} if not
}
\examples{
check_offset(offset = 0)
check_offset(10000)
}
\author{
Richèl J.C. Bilderbeek
}
