% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_b_or_t_cell_epitopes.R
\name{get_all_b_or_t_cell_epitopes}
\alias{get_all_b_or_t_cell_epitopes}
\title{Internal function}
\usage{
get_all_b_or_t_cell_epitopes(
  b_cell_or_t_cell,
  mhc_allele_names,
  verbose = FALSE
)
}
\arguments{
\item{b_cell_or_t_cell}{use "b_cell" for B cells
and "t_cell" for T cells}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
all epitopes
}
\description{
Get all epitopes from a B or T cell assay
}
\examples{
if (!beastier::is_on_ci()) {
  get_all_b_or_t_cell_epitopes(b_cell_or_t_cell = "b_cell")
  get_all_b_or_t_cell_epitopes(b_cell_or_t_cell = "t_cell")
}
}
\author{
Richèl J.C. Bilderbeek
}
