% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_mhc_ligand_epitopes.R
\name{get_all_mhc_ligand_epitopes}
\alias{get_all_mhc_ligand_epitopes}
\title{Get all epitopes from an MHC ligand assay}
\usage{
get_all_mhc_ligand_epitopes(mhc_allele_name = "all", verbose = FALSE)
}
\arguments{
\item{mhc_allele_name}{name of an MHC allele}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
all epitopes
}
\description{
Get all epitopes from an MHC ligand assay
}
\examples{
if (beastier::is_on_ci()) {
  get_all_mhc_ligand_epitopes()
  get_all_mhc_ligand_epitopes(mhc_allele_name = "cs.{HLA-A*01:01}")
}
}
\seealso{
\itemize{
\item use \link{get_all_b_cell_epitopes}
to get all epitopes from a B cell assay
\item use \link{get_all_mhc_ligand_epitopes}
to get all epitopes from an MHC ligand assay
\item use \link{get_all_t_cell_epitopes}
to get all epitopes from a T cell assay
}
}
\author{
Richèl J.C. Bilderbeek
}
