% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_iedb.R
\name{query_iedb}
\alias{query_iedb}
\title{Query IEDB}
\usage{
query_iedb(query, verbose = FALSE)
}
\arguments{
\item{query}{an IEDB query,
as created by, for example, \link{create_healthy_human_query}.
Use \link{check_query} to check if the \code{query} is valid}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
a \link[tibble]{tibble}
}
\description{
Query IEDB
}
\author{
Richèl J.C. Bilderbeek
}
