% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_iedb.R
\name{query_iedb}
\alias{query_iedb}
\title{Query the IEDB database}
\usage{
query_iedb(query, table, max_n_queries = Inf, verbose = FALSE)
}
\arguments{
\item{query}{an IEDB query,
as created by, for example, \link{create_healthy_human_query}.
Use \link{check_query} to check if the \code{query} is valid}

\item{table}{a IEDB database table,
as checked by \link{check_table}}

\item{max_n_queries}{maximum number of queries.
Use \link{Inf} to have no upper limit on this.
Lower values are useful for debugging.}

\item{verbose}{set to \link{TRUE} for more output}
}
\value{
a \link[tibble]{tibble} with as much rows as hits.
}
\description{
Query the IEDB database.
}
\note{
The IEDB returns 10k hits per query.
This means, to get the results of a query of more than 10k hits,
the IEDB needs to be queried multiple times.
\link{query_iedb} combines multiple calls to the IEDB,
using \link{query_iedb_with_offset}.
}
\author{
Richèl J.C. Bilderbeek
}
